refactor(tasks): extract TaskQueryService; registry delegates
Move the query fan-out out of TaskRegistry into a TaskQueryService that
reads the registry's providers, enabled state, and list cache, then
delegates ordering to TaskSorter. The `tasks`/`completedTasks` methods,
the `globalFanOut` helper, the ProviderFetchError typealias, and the
DST-safe Calendar.endOfToday helper all move with it.

TaskRegistry keeps thin `tasks`/`completedTasks` pass-throughs that
delegate to a lazily-constructed queryService (marked
@ObservationIgnored so the @Observable macro leaves the lazy stored
property alone). Behavior is unchanged; callers are untouched this slice.

Second slice of the #404 registry split (design doc 0005, D1 / Finding M).
